# Improve performance of information schema query.