@GuptaManan100 2022-02-05
Looking at each optimization step 1 by 1. 
For the first one, there are multiple articles describing the behaviour of `NOT EXISTS` and `LEFT JOIN ... NULL` being used as anti-joins.  This is one such [article](https://planet.mysql.com/entry/?id=24888) which also suggests that the second alternative will be faster when rows from both sides are non-null, like in this case. To make it even clearer that it is better here are the mysql explain outputs - 
```mysql
mysql [localhost:5731] {msandbox} ((none)) > explain SELECT t.table_name, t.table_type, unix_timestamp(t.create_time), t.table_comment, sum(i.file_size), sum(i.allocated_size)
    -> FROM information_schema.tables t, information_schema.innodb_sys_tablespaces i
    -> WHERE t.table_schema = database() and
    -> (i.name = concat(t.table_schema,'/',t.table_name) or i.name like concat(t.table_schema,'/',t.table_name, '#p#%'))
    -> group by t.table_name, t.table_type, unix_timestamp(t.create_time), t.table_comment, i.file_size
    -> UNION ALL
    -> SELECT table_name, table_type, unix_timestamp(create_time), table_comment, SUM( data_length + index_length), SUM( data_length + index_length)
    -> FROM information_schema.tables t
    -> WHERE table_schema = database() AND
    -> NOT EXISTS(SELECT * FROM information_schema.innodb_sys_tablespaces i WHERE i.name = concat(t.table_schema,'/',t.table_name) or i.name like concat(t.table_schema,'/',t.table_name, '#p#%'))
    -> group by table_name, table_type, unix_timestamp(create_time), table_comment;
+----+--------------------+-------+------------+------+---------------+------+---------+------+------+----------+--------------------------------------------------------------------------------------+
| id | select_type        | table | partitions | type | possible_keys | key  | key_len | ref  | rows | filtered | Extra                                                                                |
+----+--------------------+-------+------------+------+---------------+------+---------+------+------+----------+--------------------------------------------------------------------------------------+
|  1 | PRIMARY            | t     | NULL       | ALL  | NULL          | NULL | NULL    | NULL | NULL |     NULL | Using where; Open_full_table; Scanned all databases; Using temporary; Using filesort |
|  1 | PRIMARY            | i     | NULL       | ALL  | NULL          | NULL | NULL    | NULL | NULL |     NULL | Using where; Using join buffer (Block Nested Loop)                                   |
|  2 | UNION              | t     | NULL       | ALL  | NULL          | NULL | NULL    | NULL | NULL |     NULL | Using where; Open_full_table; Scanned all databases; Using temporary; Using filesort |
|  3 | DEPENDENT SUBQUERY | i     | NULL       | ALL  | NULL          | NULL | NULL    | NULL | NULL |     NULL | Using where                                                                          |
+----+--------------------+-------+------------+------+---------------+------+---------+------+------+----------+--------------------------------------------------------------------------------------+
4 rows in set, 5 warnings (0.00 sec)
```
and
```mysql
mysql [localhost:5731] {msandbox} ((none)) > explain SELECT t.table_name, t.table_type, unix_timestamp(t.create_time), t.table_comment, sum(i.file_size), sum(i.allocated_size)
    -> FROM information_schema.tables t, information_schema.innodb_sys_tablespaces i
    -> WHERE t.table_schema = database() and
    -> (i.name = concat(t.table_schema,'/',t.table_name) or i.name like concat(t.table_schema,'/',t.table_name, '#p#%'))
    -> group by t.table_name, t.table_type, unix_timestamp(t.create_time), t.table_comment, i.file_size
    -> UNION ALL
    -> SELECT table_name, table_type, unix_timestamp(create_time), table_comment, SUM( data_length + index_length), SUM( data_length + index_length)
    -> FROM information_schema.tables t
    -> LEFT OUTER JOIN information_schema.innodb_sys_tablespaces i ON i.name = concat(t.table_schema,'/',t.table_name) or i.name like concat(t.table_schema,'/',t.table_name, '#p#%')
    -> WHERE table_schema = database() AND i.name IS NULL
    -> group by table_name, table_type, unix_timestamp(create_time), table_comment;
+----+-------------+-------+------------+------+---------------+------+---------+------+------+----------+--------------------------------------------------------------------------------------+
| id | select_type | table | partitions | type | possible_keys | key  | key_len | ref  | rows | filtered | Extra                                                                                |
+----+-------------+-------+------------+------+---------------+------+---------+------+------+----------+--------------------------------------------------------------------------------------+
|  1 | PRIMARY     | t     | NULL       | ALL  | NULL          | NULL | NULL    | NULL | NULL |     NULL | Using where; Open_full_table; Scanned all databases; Using temporary; Using filesort |
|  1 | PRIMARY     | i     | NULL       | ALL  | NULL          | NULL | NULL    | NULL | NULL |     NULL | Using where; Using join buffer (Block Nested Loop)                                   |
|  2 | UNION       | t     | NULL       | ALL  | NULL          | NULL | NULL    | NULL | NULL |     NULL | Using where; Open_full_table; Scanned all databases; Using temporary; Using filesort |
|  2 | UNION       | i     | NULL       | ALL  | NULL          | NULL | NULL    | NULL | NULL |     NULL | Using where; Not exists; Using join buffer (Block Nested Loop)                       |
+----+-------------+-------+------------+------+---------------+------+---------+------+------+----------+--------------------------------------------------------------------------------------+
4 rows in set, 1 warning (0.00 sec)
```

A dependent subquery is being converted to a block nested loop join.