@GuptaManan100 2022-02-05
For the second optimization - 
Let's call this predicate `(i.name = concat(t.table_schema,'/',t.table_name) or i.name like concat(t.table_schema,'/',t.table_name, '#p#%'))` as `A`;
If a row satisfies predicate A, then `i.name` will be NOT NULL.
The first half of the union is doing a cross join and taking all rows which match the predicate A. It is safe to convert the cross join to left join on predicate A with an additional check that `i.name` is NOT NULL so that no additional rows from the left join qualify

Following this, we now have 2 parts of a UNION, both having a left join on predicate A, the first one selecting rows with `i.name` NOT NULL and the second selecting rows with `i.name` is NULL. 

As long as we don't have rows that qualify from both the sides having the same values for the columns in the group by part, this rewrite should be okay.